backend/requirements/develop.txt
appdirs==1.4.4; python_full_version >= "3.6.2"
appnope==0.1.2; sys_platform == "darwin" and python_version >= "3.7"
asgiref==3.4.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
astroid==2.6.5; python_version >= "3.6" and python_version < "4.0"
atomicwrites==1.4.0; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.4.0"
attrs==21.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
backcall==0.2.0; python_version >= "3.7"
backports.entry-points-selectable==1.1.0; python_version >= "2.7" and python_full_version >= "3.6.1"
black==21.7b0; python_full_version >= "3.6.2"
certifi==2021.5.30; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
cffi==1.14.6; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
cfgv==3.3.0; python_full_version >= "3.6.1"
charset-normalizer==2.0.4; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3"
click==8.0.1; python_version >= "3.6" and python_full_version >= "3.6.2"
colorama==0.4.4; python_version >= "3.7" and python_version < "4.0" and sys_platform == "win32" and python_full_version >= "3.6.2" and platform_system == "Windows" and (python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0")
coreapi==2.3.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
coreschema==0.0.4; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
coverage==5.5; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
cryptography==3.4.7; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
decorator==5.0.9; python_version >= "3.7"
defusedxml==0.7.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
distlib==0.3.2; python_full_version >= "3.6.1"
django-cors-headers==3.7.0; python_version >= "3.6"
django-coverage-plugin==2.0.0
django-debug-toolbar-request-history==0.1.3
django-debug-toolbar==3.2.1; python_version >= "3.6"
django-extensions==3.1.3; python_version >= "3.6"
django-filter==2.4.0; python_version >= "3.5"
django-templated-mail==1.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
django==3.2.5; python_version >= "3.6"
djangorestframework-camel-case==1.2.0; python_version >= "3.5"
djangorestframework-simplejwt==4.7.2; python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
djoser==2.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
factory-boy==3.2.0; python_version >= "3.6"
faker==8.10.3; python_version >= "3.6"
filelock==3.0.12; python_full_version >= "3.6.1"
flake8==3.9.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
gunicorn==20.1.0; python_version >= "3.5"
identify==2.2.11; python_full_version >= "3.6.1"
idna==3.2; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.5"
iniconfig==1.1.1; python_version >= "3.6"
ipython-genutils==0.2.0; python_version >= "3.7"
ipython==7.26.0; python_version >= "3.7"
isort==5.9.2; python_full_version >= "3.6.1" and python_version < "4.0"
itypes==1.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
jedi==0.18.0; python_version >= "3.7"
jinja2==3.0.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
lazy-object-proxy==1.6.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.6.0"
markupsafe==2.0.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
matplotlib-inline==0.1.2; python_version >= "3.7"
mccabe==0.6.1; python_version >= "3.6" and python_version < "4.0" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0")
mypy-extensions==0.4.3; python_full_version >= "3.6.2"
nodeenv==1.6.0; python_full_version >= "3.6.1"
oauthlib==3.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
packaging==21.0; python_version >= "3.6"
parso==0.8.2; python_version >= "3.7"
pathspec==0.9.0; python_full_version >= "3.6.2"
pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.7"
pickleshare==0.7.5; python_version >= "3.7"
pillow==8.3.1; python_version >= "3.6"
platformdirs==2.0.2; python_full_version >= "3.6.1"
pluggy==0.13.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pre-commit==2.13.0; python_full_version >= "3.6.1"
prompt-toolkit==3.0.19; python_full_version >= "3.6.1" and python_version >= "3.7"
psycopg2-binary==2.9.1; python_version >= "3.6"
ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.7"
py==1.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pycodestyle==2.7.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pycparser==2.20; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pyflakes==2.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pygments==2.9.0; python_version >= "3.7"
pyjwt==2.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
pylint==2.9.5; python_version >= "3.6" and python_version < "4.0"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
pytest-django==4.4.0; python_version >= "3.5"
pytest==6.2.4; python_version >= "3.6"
python-dateutil==2.8.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python-environ==0.4.54
python3-openid==3.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pytz==2021.1; python_version >= "3.6"
pyyaml==5.4.1; python_full_version >= "3.6.1"
regex==2021.7.6; python_full_version >= "3.6.2"
requests-oauthlib==1.3.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
requests==2.26.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
six==1.16.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6")
social-auth-app-django==4.0.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-core==4.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
sqlparse==0.4.1; python_version >= "3.6"
text-unidecode==1.3; python_version >= "3.6"
toml==0.10.2; python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.6.1" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6")
tomli==1.0.4; python_version >= "3.6" and python_full_version >= "3.6.2"
traitlets==5.0.5; python_version >= "3.7"
uritemplate==3.0.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
urllib3==1.26.6; python_full_version >= "3.6.1" and python_version < "4" and python_full_version < "4.0.0"
virtualenv==20.6.0; python_full_version >= "3.6.1"
wcwidth==0.2.5; python_full_version >= "3.6.1" and python_version >= "3.7"
wrapt==1.12.1; python_version >= "3.6" and python_version < "4.0"
